In this episode of "Doesn't this guy have something better to do?" we'll be enjoying the following:
Fixed a "bug" in the model where the activeFIFOIndex of an input channel is -1 when an output channel is in state 6. This was caused by the ModelRunner generating messages FROM random nodes TO random nodes instead of just from 0 fo wherever.

On the UI side of things there are lots of improvements and a known bug:
 1. The step-by-step simulation can now be restarted after it finishes
 2. The sent/received packets are now displayed in a ScrollPane, giving us the option to send an unlimited* amount of packets
 3. Some other minor changes also took place.

 I'm about to embark on a vary adventurous journey depicting the graphs of input/output channels for a node at a certain point in time. Cool stuff.

 * Well... at least until it throws an exception somewhere